+++ source/lib/sysdep/os/win/wposix/wtime.cpp
@@ -263,7 +263,7 @@ match_string (const char **buf, const char **strs)
 	int i = 0;
 
 	for (i = 0; strs[i] != NULL; ++i) {
-		int len = strlen (strs[i]);
+		size_t len = strlen (strs[i]);
 
 		if (strncasecmp (*buf, strs[i], len) == 0) {
 			*buf += len;
